[compat, modify] convert enum to enum class for PreFsmActionListenerType

+++ src/lib/rtm/RTObject.cpp
@@ -2490,7 +2490,7 @@ namespace RTC
                           bool autoclean)
   {
     m_fsmActionListeners.
-      preaction_[listener_type].addListener(listener, autoclean);
+      preaction_[static_cast<uint8_t>(listener_type)].addListener(listener, autoclean);
   }
   
   
@@ -2508,7 +2508,7 @@ namespace RTC
                              PreFsmActionListener* listener)
   {
     m_fsmActionListeners.
-      preaction_[listener_type].removeListener(listener);
+      preaction_[static_cast<uint8_t>(listener_type)].removeListener(listener);
   }
